Top level components layout reforged; On start menu; undefined supported in context
