+++ src/components/ChordsEditor/ChordsEditor.scss
@@ -1,0 +1,8 @@
+@import '@stylesheets/common.scss';
+
+.chords-editor__text {
+    margin-left: calc(50% - 20rem);
+    padding-top: $spacing-xl;
+    margin-bottom: 75px;
+}
+
