+++ src/components/ChordsEditor/InstrumentalPart/ChordsLine.scss
