+++ src/components/ChordsEditor/InstrumentalPart/InstrumentalPart.tsx
