+++ src/components/ChordsEditor/InstrumentalPart/index.ts
