+++ src/components/ChordsEditor/LyricsPart/Line/LyricsLine/LyricsLine.scss
