+++ src/components/ChordsEditor/LyricsPart/Line/LyricsLine/index.ts
