+++ src/components/ChordsEditor/LyricsPart/Line/LyricsLine/utils.ts
