+++ src/components/ChordsEditor/LyricsPart/LyricsPart.scss
