+++ src/components/ChordsEditor/LyricsPart/index.ts
