+++ src/components/ChordsEditor/LyricsPart/utils.ts
