+++ src/components/ChordsEditor/SongPart/PartHeader.scss
