+++ src/components/ChordsEditor/SongPart/SongPart.scss
