+++ src/components/ChordsEditor/SongPart/SongPart.tsx
