+++ src/components/ChordsEditor/SongPart/index.ts
