+++ src/components/ChordsEditor/SongPart/types.ts
