+++ src/components/ChordsEditor/index.ts
