--- src/components/Editor/EditorPage.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import { Container } from 'rsuite';
-import { ChordsEditor } from './ChordsEditor';
-
-export const EditorPage: React.FC = () => {
-    return (
-        <Container>
-            <ChordsEditor />
-        </Container>
-    );
-};
