--- src/components/Editor/chordsEditor/ChordsEditor.scss
@@ -1,12 +1,0 @@
-@import '@stylesheets/common.scss';
-
-.chords-editor {
-    height: calc(100vh - 84.5px);
-    user-select: none;
-
-    &__text {
-        margin-left: calc(50% - 20rem);
-        padding-top: $spacing-xl;
-        margin-bottom: 75px;
-    }
-}
