--- src/components/Editor/chordsEditor/utils/index.ts
@@ -1,3 +1,0 @@
-export * from './checkFormat';
-export * from './utils';
-export * from './getMonospacedText';
