--- src/components/Editor/index.ts
@@ -1,1 +1,0 @@
-export * from './EditorPage';
