+++ src/components/Home/Home.scss
@@ -1,0 +1,7 @@
+@import '@stylesheets/common.scss';
+
+.home-page-content {
+    height: calc(100vh - 84.5px);
+    user-select: none;
+}
+
