+++ src/components/MonospacedModal/MonospacedModal.scss
