+++ src/utils/checkFormat.ts
