+++ src/utils/getNewSong.ts
@@ -3,5 +3,9 @@ import { Song } from '@model/song';
 export const getNewSong = (): Song => ({
     title: '',
     author: '',
-    songBody: [],
+    songBody: [
+        {
+            lines: [],
+        },
+    ],
 });
